+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix crash when entering mba-testing for the first time
 - Fix pvp points desyncing on remote
 - Fix players being unaffected by time scales
+- Fix invalid party handling on player map change
 
 ## [0.8.2] 2026-07-26